refactor(quest): 詳細ページのリファクタリング
