+++ feature/quest/lib/feature_quest.dart
@@ -1,6 +1,5 @@
 export 'src/ui/page/add/quest_add_page.dart';
 
 export 'src/ui/page/detail/quest_detail_page.dart';
-export 'src/ui/page/detail/quest_detail_route.dart';
 
 export 'src/ui/page/list/quest_list_page.dart';
